Move FrozenObject to nengo.base
