+++ nengo/neurons.py
@@ -2,8 +2,9 @@ import warnings
 
 import numpy as np
 
+from nengo.base import FrozenObject
 from nengo.exceptions import SimulationError, ValidationError
-from nengo.params import Parameter, NumberParam, FrozenObject
+from nengo.params import Parameter, NumberParam
 
 
 def settled_firingrate(step_math, J, states, dt=0.001, settle_time=0.1, sim_time=1.0):
