+++ nengo/params.py
@@ -582,99 +582,3 @@ class FunctionParam(Parameter):
             )
 
         return function_info
-
-
-class FrozenObject:
-    """An object with parameters that cannot change value after instantiation.
-
-    Since such objects are read-only ("frozen"), they can be safely used in
-    multiple locations, compared, etc.
-    """
-
-    # Order in which parameters have to be initialized.
-    # Missing parameters will be initialized last in an undefined order.
-    # This is needed for pickling and copying of Nengo objects when the
-    # parameter initialization order matters.
-    _param_init_order = []
-
-    def __init__(self):
-        self._paramdict = collections.OrderedDict(
-            (k, v)
-            for k, v in inspect.getmembers(type(self))
-            if isinstance(v, Parameter) and not isinstance(v, ObsoleteParam)
-        )
-        for p in self._params:
-            if not p.readonly:
-                msg = "All parameters of a FrozenObject must be readonly"
-                raise ReadonlyError(attr=p, obj=self, msg=msg)
-        self.__argreprs = None
-
-    @property
-    def _params(self):
-        return list(self._paramdict.values())
-
-    def __eq__(self, other):
-        if self is other:  # quick check for speed
-            return True
-        return type(self) == type(other) and all(
-            p.equal(self, other) for p in self._params
-        )
-
-    def __hash__(self):
-        return hash((type(self), tuple(p.hashvalue(self) for p in self._params)))
-
-    def __getstate__(self):
-        d = dict(self.__dict__)
-        d.pop("_paramdict")  # do not pickle the param dict itself
-        for k in self._paramdict:
-            d[k] = getattr(self, k)
-
-        return d
-
-    def __setstate__(self, state):
-        FrozenObject.__init__(self)  # set up the param dict
-
-        for attr in self._param_init_order:
-            setattr(self, attr, state.pop(attr))
-        for attr in set(self._paramdict).difference(self._param_init_order):
-            setattr(self, attr, state.pop(attr))
-
-        self.__dict__.update(state)
-
-    def __repr__(self):
-        if isinstance(self._argreprs, str):
-            return "<%s at 0x%x>" % (type(self).__name__, id(self))
-        return "%s(%s)" % (type(self).__name__, ", ".join(self._argreprs))
-
-    @property
-    def _argreprs(self):
-        if self.__argreprs is not None:
-            return self.__argreprs
-
-        # get arguments to display from __init__ functions
-        spec = inspect.getfullargspec(type(self).__init__)
-        defaults = {}
-        if spec.defaults is not None:
-            defaults.update(zip(spec.args[-len(spec.defaults) :], spec.defaults))
-
-        self.__argreprs = []
-        for arg in spec.args[1:]:  # start at 1 to drop `self`
-            if not hasattr(self, arg):
-                # We rely on storing the initial arguments. If we don't have
-                # them, we don't auto-generate a repr.
-                self.__argreprs = "Cannot find %r" % arg
-                break
-            value = getattr(self, arg)
-
-            param = self._paramdict.get(arg, None)
-            if arg in defaults and defaults[arg] is not Default:
-                not_default = not equal(value, defaults[arg])
-            elif param is not None and param.default is not Unconfigurable:
-                not_default = not equal(value, param.default)
-            else:
-                not_default = True
-
-            if not_default:
-                self.__argreprs.append("%s=%r" % (arg, value))
-
-        return self.__argreprs
